Improvement and bugs fix
Dear NBODY6 users,

There have been several developments since the new version was announced
on Friday 13th (which proved an unlucky day for me). A further
complication was that the "official" GPU computer has had a disk crash
and sometimes does not work at all. However, a second GPU machine with
almost an identical code is performing test runs. In addition, Jarrod is
working with the newly downloaded code and has spotted several problems
which were fixed. A fresh tar file has been created but is still not on
Github. It may therefore be useful for some to receive directly the
modified routines. This is applicable to the previous and current
versions.

absorb.f:  A small safety condition included for rare case (dir Nchain
only).

expel.f:   Procedures for standard common envelope have been separated
from
optional tidal disruption by black holes (dir Block & Ncode.)
An exceptional case of two black holes also needed to be defined
as coalescence (without PN).

expel2.f:  Small change to allow velocity kick for WDs with option #25 >
0.
Current version would give a kick regardless. (only dir ARint.)

impact.f:  A small but important bug fix (dir Block only) which
initializes
the indicator IPHASE for use by algorithmic regularization and
compact unperturbed 3/4-body systems. Although this only involved
activating a suppressed statement, it took two days to unravel.

unpert.f:  The setting of an index at the end is wrong and should be
deleted.
Only for replacing in dir Block.

ksapo.f:   A NaN appeared on some occasions, leading to infinite
looping.
Fortunately, these conditions could be reproduced most times but
including diagnostics in the offending routine produced different
outcome. Eventually a clue was obtained from a write in the middle
rather than the end (which often works). This routine is called
with three arguments and the latest version used a variable not
initialized in one of the cases. The same version should go in
rather than the end (which often works). This routine is called
with three arguments and the latest version used a variable not
initialized in one of the cases. The same routine should go in
both relevant directories (Ncode & Block).
As recently as today it appears that calls with a negative
argument (from expel.f) did not include a factor of 0.5 which is
needed in two-body regularization (probably rarely used).

ksterm.f:  The algorithmic decision-making is actually wrong. This led
to a
strange behaviour which was hard to resolve. All that was needed
was to delete one IF statement. Both directories are affected but
slightly different (noted in the Subject as ksterm.f and
ksterm.f.blk, respectively).

newton.f:  Somewhat surprisingly the Newton-Raphson iteration failed (no
convergence in 15 attempts). Investigation of the reproduced
condition finally revealed the cause. Namely the first guess
was so good that the derivative was factor 10 smaller than the
function and consequently the second guess became very large,
leading to too slow convergence. A cheap fix limiting the change
to 10 % (both signs) solved the problem. This routine to be
copied to dir Block.

chlist.f:  Here is a new algorithm to celebrate notice #150 since 2007.
Until now the perturbation list for chain/ARchain have been based
on determining the maximum value of m_p/r_p^3 regardless of the
internal values of masses and separation. These days it is usual
to employ rather large mass ratios. It would therefore seem more
appropriate to evaluate the relative perturbation acting on each
chain distance, This can be done quite fast using the same value
for the perturber distance (squared and cubed). All that is
required is to determine an appropriate perturber list rather
than actual values.
The new procedure may appear to be quite similar. However, in the
case of different chain masses, the relative perturbation on one
large chain distance with small masses would be significantly
larger than the old determination and hence more sensitive for
the decision-making. The two versions only differ in their common
declarations so will be denoted chlist1.f and chlist2.f in the
Subject (dir Nchain & ARint,respectively).

impact.f:  The transition to the Valtonen three-body stability criterion
resulted in the failure of some conditions. This was caused by
the optional generation of diagnostics which has now been fixed.
Versions for dirs Ncode and Block are indicated in the Subject.

ksint.f:   The case of a strong binary-binary interaction can be treated
by chain regularization for separations which may otherwise
cause numerical problems. This can most simply be achieved by
modifying an acceptance criterion if the outer eccentricity
exceeds unity (i.e. hyperbolic encounter). The basic criterion
for chain treatment (GAMMA > 0.05) ensures the encounter is
sufficiently critical. This also applies to the case when the
perturber is a binary. Two versions are quoted in the Subject.

ksintp.f:  The parallel routine belongs in dir Block.

chain.f:   Decision-making has been improved (dir ARchain).

ksrect.f:  At the last moment I was hit by an extremely rare and fatal
condition. This occurred when the eccentricity calculated from
well behaved variables became 1.0 when printed with extra
precision to 8 figures which gave rise to a NaN. This might
have been caused by a Kozai cycle. After some experimentation
the fix was easy and a skip was accepted after lowering the
limit to 0.999999. For safety this routine should be included
in dirs Ncode and Block.